+++ ReleaseNotes.md
@@ -1,2 +1,3 @@
 # Release 2.7.4
+* Adds the `TryReadQueryAsync` database method
 * Adds the `ReadQueryWhenReadyAsync` database method